move customer creation to main class
